osd: copy the reqids even if the object is deleted during promotion
If the object is deleted on the base tier, and the reqids are not copied
during promotion, this again leads to the 'ops not idempotent' problem.
This fix moves the 'object not existing' detection to do_osd_ops. And
for the copy-get op, copies the reqids even if the object doesn't exist.

Signed-off-by: Zhiqiang Wang <zhiqiang.wang@intel.com>